# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Application API documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/companies": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Get all companies",
                "description": "Get all companies, this route worked only for admin role.",
                "operationId": "getCompanies",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/company/{company_id}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Get company by ID",
                "description": "Get one company with ID, this route worked only for admin role.",
                "operationId": "getCompany",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/company/create": {
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "Create company",
                "description": "Create company, this route worked only for admin role.",
                "operationId": "addCompany",
                "requestBody": {
                    "description": "Create company",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "company_name",
                                    "supervisor",
                                    "address",
                                    "email",
                                    "website",
                                    "phone",
                                    "manager_id"
                                ],
                                "properties": {
                                    "company_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Company name"
                                    },
                                    "supervisor": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Oliver Jake "
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Uzbekistan Tashkent region, Parkent town"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "test@gmail.com"
                                    },
                                    "website": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "https://www.google.com/"
                                    },
                                    "phone": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "3333333333"
                                    },
                                    "manager_id": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/company/update/{company_id}": {
            "put": {
                "tags": [
                    "Company"
                ],
                "summary": "Get company by ID",
                "description": "Update company with ID, this route worked only for admin role.",
                "operationId": "updateCompany",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update company",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "company_name",
                                    "supervisor",
                                    "address",
                                    "email",
                                    "website",
                                    "phone",
                                    "manager_id"
                                ],
                                "properties": {
                                    "company_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Company name"
                                    },
                                    "supervisor": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Oliver Jake "
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Uzbekistan Tashkent region, Parkent town"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "test@gmail.com"
                                    },
                                    "website": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "https://www.google.com/"
                                    },
                                    "phone": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "3333333333"
                                    },
                                    "manager_id": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/company/delete/{company_id}": {
            "delete": {
                "tags": [
                    "Company"
                ],
                "summary": "Delete company by ID",
                "description": "Delete company with ID, this route worked only for admin role.",
                "operationId": "deleteCompany",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Company": {
                "title": "Company",
                "description": "Company model, companies can be changed or added only by admins.",
                "xml": {
                    "name": "Company"
                }
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}